Fixing return value of one of the syscalls

+++ include/syscall.h
@@ -93,7 +93,7 @@ int syscall_fentry(char * s, uint16_t entry);
 int syscall_pspawn(int pd, char ** argv, size_t argc);
 int syscall_pexec(int pd, char ** argv, size_t argc);
 int syscall_pload(const char * filename);
-int syscall_pstate(int pid);
+int8_t syscall_pstate(int pid);
 void syscall_pexit(int pid);
 int syscall_pexitcode(int pid);
